move status commands to their own package, and move component registration to that package
