+++ cmd/juju/commands/common.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"net/http"
 	"path"
-	"time"
 
 	"github.com/juju/cmd"
 	"github.com/juju/errors"
@@ -258,23 +257,4 @@ func (c *csClient) authorize(curl *charm.URL) (*macaroon.Macaroon, error) {
 		return nil, errors.Trace(err)
 	}
 	return m, nil
-}
-
-// formatStatusTime returns a string with the local time
-// formatted in an arbitrary format used for status or
-// and localized tz or in utc timezone and format RFC3339
-// if u is specified.
-func formatStatusTime(t *time.Time, formatISO bool) string {
-	if formatISO {
-		// If requested, use ISO time format.
-		// The format we use is RFC3339 without the "T". From the spec:
-		// NOTE: ISO 8601 defines date and time separated by "T".
-		// Applications using this syntax may choose, for the sake of
-		// readability, to specify a full-date and full-time separated by
-		// (say) a space character.
-		return t.UTC().Format("2006-01-02 15:04:05Z")
-	} else {
-		// Otherwise use local time.
-		return t.Local().Format("02 Jan 2006 15:04:05Z07:00")
-	}
 }
